moved regEx library from cms-cli to cms-lib
